added save load tick and cross icons, added cancel option to saving and loading, fixed issue where buttons could be used in menus, made x and y range global, renamed file type from root to wdld

+++ Assets/Scripts/UI/RerollPanel.cs
@@ -21,33 +21,48 @@ public class RerollPanel : MonoBehaviour
     
     public void RerollWoodland()
     {
-        worldState.DeleteAllClearings();
-        worldState.DeleteAllClearings();
-        mapGenerator.GenerateClearings();
-        riverGenerator.GenerateRiver();
-        mapGenerator.GeneratePaths();
-        clearingInfoGenerator.GenerateDenizens();
-        clearingInfoGenerator.GenerateClearingNames();
+        if (worldState.menuState == MenuState.Default)
+        {
+            worldState.DeleteAllClearings();
+            worldState.DeleteAllClearings();
+            mapGenerator.GenerateClearings();
+            riverGenerator.GenerateRiver();
+            mapGenerator.GeneratePaths();
+            clearingInfoGenerator.GenerateDenizens();
+            clearingInfoGenerator.GenerateClearingNames();
+        }
     }
 
     public void RerollPaths()
     {
-        worldState.DeleteAllPaths();
-        mapGenerator.GeneratePaths();
+        if (worldState.menuState == MenuState.Default)
+        {
+            worldState.DeleteAllPaths();
+            mapGenerator.GeneratePaths();
+        }
     }
 
     public void RerollRiver()
     {
-        riverGenerator.GenerateRiver();
+        if (worldState.menuState == MenuState.Default)
+        {
+            riverGenerator.GenerateRiver();
+        }
     }
 
     public void RerollDenizens()
     {
-        clearingInfoGenerator.GenerateDenizens();
+        if (worldState.menuState == MenuState.Default)
+        {
+            clearingInfoGenerator.GenerateDenizens();
+        }
     }
 
     public void RerollNames()
     {
-        clearingInfoGenerator.GenerateClearingNames();
+        if (worldState.menuState == MenuState.Default)
+        {
+            clearingInfoGenerator.GenerateClearingNames();
+        }
     }
 }
